Changes logging_module name to logging and refactors relative files
