+++ mipengine/node/monetdb_interface/monet_db_connection.py
@@ -4,7 +4,7 @@ from typing import List
 import pymonetdb
 
 from mipengine.node import config as node_config
-from mipengine.node import logging_module as logging
+from mipengine.node import logging as logging
 
 BROKEN_PIPE_MAX_ATTEMPTS = 50
 OCC_MAX_ATTEMPTS = 50
