+++ mipengine/node/tasks/merge_tables.py
@@ -8,7 +8,7 @@ from mipengine.node.monetdb_interface import common_actions
 from mipengine.node.monetdb_interface import merge_tables
 from mipengine.node.monetdb_interface.common_actions import create_table_name
 from mipengine.node.monetdb_interface.merge_tables import validate_tables_can_be_merged
-from mipengine.node.logging_module import log_method_call
+from mipengine.node.logging import log_method_call
 
 
 @shared_task
