+++ mipengine/node/tasks/remote_tables.py
@@ -4,7 +4,7 @@ from celery import shared_task
 
 from mipengine.node_tasks_DTOs import TableInfo
 from mipengine.node.monetdb_interface import remote_tables
-from mipengine.node.logging_module import log_method_call
+from mipengine.node.logging import log_method_call
 
 
 @shared_task
